backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.13",
    "langchain-experimental>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "pgvector>=0.4.2",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
